+++ Services/AssessmentQuestion/src/Authoring/DomainModel/Question/Answer/Option/Value/AnswerOptionValueIsUserdefined.php
@@ -1,0 +1,7 @@
+<?php
+namespace ILIAS\AssessmentQuestion\Authoring\DomainModel\Question\Answer\Option\Value;
+
+
+interface AnswerOptionValueIsUserdefined {
+
+}