+++ Services/AssessmentQuestion/CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 03.07.2019
 - MST: globale Struktur src/Data/Domain verschoben nach Servicec/AssessmentQuestion/src/Common - Muss nun zusätzlich geprüft und fertiggestellt werden. Readme (Services/AssessmentQuestion/src/Common/DomainModel/Aggregate/docs/documentation/Readme.md) muss fertiggestellt werden.
 - MST: Noch einige zusätzliche Klassen bei Event hinzugefügt
+
+## 08.07.2019
+Moddeling Answer-Entity
+Usage see
+-> Services/AssessmentQuestion/src/Authoring/DomainModel/Question/Answer/Option/AnswerOptionFactory.php